move jmh to test
