+++ flink/flink-1.14/pom.xml
@@ -114,11 +114,13 @@
       <groupId>org.openjdk.jmh</groupId>
       <artifactId>jmh-core</artifactId>
       <version>${dep.jmh.version}</version>
+      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.openjdk.jmh</groupId>
       <artifactId>jmh-generator-annprocess</artifactId>
       <version>${dep.jmh.version}</version>
+      <scope>test</scope>
     </dependency>
   </dependencies>
