--- flink/flink-1.14/src/jmh/README.md
@@ -1,4 +1,0 @@
-To run jmh, you need to:
-
-1. just execute main in benchmark.java
-2. sometimes you need to set <flink.scope> to compile in pom.xml to run with flink dependencies
