+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/jmh/README.md
@@ -1,0 +1,3 @@
+Java Microbenchmark Harness or JMH is a tool for creating Java microbenchmarks. 
+
+To run jmh, just change the @Ignore to @Test in the test.
